requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2

# LangChain and related
langchain>=0.0.350
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Vector store and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Image processing
Pillow>=10.1.0

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2

# Utilities
python-multipart>=0.0.6
typing-extensions>=4.8.0
uuid>=1.30
pathlib>=1.0.1
timm

# Development dependencies
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
